# robin/openapi.yaml
openapi: "3.1.0"
info:
  title: Robin API
  version: "0.3.0"
  description: REST API for Robin.OS — a personal knowledge management system.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /entries:
    post:
      operationId: createEntry
      summary: Create a new entry (queues async processing)
      tags: [Entries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createEntryBodySchema"
      responses:
        "202":
          description: Entry created and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/entryCreatedResponseSchema"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
    get:
      operationId: listEntries
      summary: List recent entries
      tags: [Entries]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/entryListResponseSchema"
  /entries/{id}:
    get:
      operationId: getEntry
      summary: Get an entry by ID
      tags: [Entries]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: The entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/entryResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /entries/{id}/fragments:
    get:
      operationId: listEntryFragments
      summary: Get all fragments derived from an entry
      tags: [Entries]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Fragments for this entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fragmentListResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"

  # ── Fragments ──────────────────────────────────────────────────────
...
  /api/content/{type}/{key}:
    get:
      operationId: getContent
      summary: Read raw or structured content
      tags: [Content]
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [fragment, entry, thread, person]
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content (raw or structured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/contentRawResponseSchema"
        "400":
          description: Invalid type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
    put:
      operationId: updateContent
      summary: Write structured content
      tags: [Content]
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [fragment, entry, thread, person]
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/okResponseSchema"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"

  # ── Robin (Config) ─────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by better-auth after sign-in
    hmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature of request body
    mcpToken:
      type: apiKey
      in: query
      name: token
      description: JWT token for MCP access

  # ── Shared Parameters ──────────────────────────────────────────────
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Resource ID or lookup key
      schema:
        type: string
...
  schemas:
...
    errorResponseSchema:
      type: object
      properties:
        error:
          type: string
        fields: {}
      required: [error]
      additionalProperties: false
    okResponseSchema:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
      required: [ok]
      additionalProperties: false
...
    createEntryBodySchema:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        title:
          type: string
        source:
          type: string
          default: api
        type:
          type: string
          default: thought
        vaultId:
          type: string
      required: [content]
      additionalProperties: false
...
    entryResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
        source:
          type: string
        vaultId:
          type: string
          nullable: true
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, lookupKey, userId, slug, title, content, type, source, vaultId, state, repoPath, createdAt, updatedAt]
      additionalProperties: false
    entryCreatedResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
        source:
          type: string
        vaultId:
          type: string
          nullable: true
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        jobId:
          type: string
        status:
          type: string
          enum: [queued]
      required: [id, lookupKey, userId, slug, title, content, type, source, vaultId, state, repoPath, createdAt, updatedAt, jobId, status]
      additionalProperties: false
    entryListResponseSchema:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/entryResponseSchema"
      required: [entries]
      additionalProperties: false

    # ── Fragments ──────────────────────────────────────────────────
...
    fragmentResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        title:
          type: string
        type:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        entryId:
          type: string
          nullable: true
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, lookupKey, userId, slug, title, type, tags, entryId, state, repoPath, createdAt, updatedAt]
      additionalProperties: false
...
    fragmentListResponseSchema:
      type: object
      properties:
        fragments:
          type: array
          items:
            $ref: "#/components/schemas/fragmentResponseSchema"
      required: [fragments]
      additionalProperties: false

    # ── Threads ────────────────────────────────────────────────────
...
    contentRawResponseSchema:
      type: object
      properties:
        content:
          type: string
      required: [content]
      additionalProperties: false
...